.old/ifhai/node/node.py
from functools import reduce
import random
from abc import ABC, abstractmethod
import asyncio
import threading
from typing import Callable, Generic, TypeVar, Any

import tree
from networkx.classes.multidigraph import MultiDiGraph
import networkx as nx

from ifhai.utils.parellelization.thread_safety import ThreadSafe
